requirements-step4.txt
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.21.0
paho-mqtt>=1.6.0